Extract color handling into its own module
